+++ src/main/java/kr/co/suitcarrier/web/entity/post/PostImage.java
@@ -1,12 +1,12 @@
-package kr.co.suitcarrier.web.entity;
+package kr.co.suitcarrier.web.entity.post;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 
 @Entity
 @Getter
-@Table(name = "rental_date")
-public class RentalDate {
+@Table(name = "post_image")
+public class PostImage {
     @Id
     @Column(name="id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -16,9 +16,9 @@ public class RentalDate {
     @JoinColumn(name = "post_id", referencedColumnName = "id")
     private Post post;
 
-    @Column(name="start_at")
-    private String startAt;
+    @Column(name="thumnail", nullable = false)
+    private String thumnail;
 
-    @Column(name="end_at")
-    private String endAt;
+    @Column(name="image_url", nullable = false)
+    private String imageUrl;
 }
