+++ src/main/java/kr/co/suitcarrier/web/entity/post/RentalDate.java
@@ -1,12 +1,13 @@
-package kr.co.suitcarrier.web.entity;
+package kr.co.suitcarrier.web.entity.post;
 
 import jakarta.persistence.*;
+import kr.co.suitcarrier.web.entity.post.Post;
 import lombok.Getter;
 
 @Entity
 @Getter
-@Table(name = "post_image")
-public class PostImage {
+@Table(name = "rental_date")
+public class RentalDate {
     @Id
     @Column(name="id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -16,9 +17,9 @@ public class PostImage {
     @JoinColumn(name = "post_id", referencedColumnName = "id")
     private Post post;
 
-    @Column(name="thumnail", nullable = false)
-    private String thumnail;
+    @Column(name="start_at")
+    private String startAt;
 
-    @Column(name="image_url", nullable = false)
-    private String imageUrl;
+    @Column(name="end_at")
+    private String endAt;
 }
