--- src/main/java/kr/co/suitcarrier/web/repository/PostRepository.java
@@ -1,11 +1,0 @@
-package kr.co.suitcarrier.web.repository;
-
-import kr.co.suitcarrier.web.entity.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
-
-
-}
