+++ src/main/java/kr/co/suitcarrier/web/repository/post/PostImageRepository.java
@@ -1,6 +1,6 @@
-package kr.co.suitcarrier.web.repository;
+package kr.co.suitcarrier.web.repository.post;
 
-import kr.co.suitcarrier.web.entity.PostImage;
+import kr.co.suitcarrier.web.entity.post.PostImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
