+++ src/main/java/kr/co/suitcarrier/web/repository/post/PostStateRepository.java
@@ -1,6 +1,6 @@
-package kr.co.suitcarrier.web.repository;
+package kr.co.suitcarrier.web.repository.post;
 
-import kr.co.suitcarrier.web.entity.PostState;
+import kr.co.suitcarrier.web.entity.post.PostState;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
