+++ src/main/java/kr/co/suitcarrier/web/repository/post/ProductRepository.java
@@ -1,6 +1,7 @@
-package kr.co.suitcarrier.web.repository;
+package kr.co.suitcarrier.web.repository.post;
 
-import kr.co.suitcarrier.web.entity.Product;
+
+import kr.co.suitcarrier.web.entity.post.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
